frontend/requirements.txt
blinker==1.7.0
click==8.1.7
flask==3.0.2
importlib-metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
kafka-python==2.0.2
MarkupSafe==2.1.5
python-dotenv==1.0.1
werkzeug==3.0.2
zipp==3.18.1
